src/third_party/blink/web_tests/external/wpt/service-workers/service-worker/fetch-request-css-base-url.https-expected.txt

This is a testharness.js-based test.
PASS global setup
FAIL base URL when service worker does respondWith(fetch(responseUrl)). assert_equals: referrer expected "https://web-platform.test:8444/service-workers/service-worker/resources/fetch-request-css-base-url-style.css?fetch" but got "https://web-platform.test:8444/service-workers/service-worker/resources/fetch-request-css-base-url-style.css"
PASS base URL when service worker does respondWith(new Response()).
PASS cleanup global state
Harness: the test ran to completion.
